Move Property to own file
Since Property (via Params) is shared by both Project and Build,
it was removed from project.go and given its own file.
